Add ability to SSH into the agents node
This is very useful for debugging.

I have opened the SSH port only to the subnet, and not to the Internet.
That is clearly for extra security.

That means you will need to use the master node (which is publicly accessible, but IP-restricted)
as a "proxy" in order to SSH into the agents node from your laptop.
I have included documentation that explains how to do that.