Fix docker healthcheck on jenkins master
The below command fails in the situation where a connection succeeds and it gets a response >400 which occurs when github auth is enabled and a 403 is returned indicating authentication is required.

`CMD curl -f http://127.0.0.1:8080 || exit 1`

This change merely checks whether a connection is possible regardless of the http responses.

Solo: @smford