Force bucket creation to fail if the team_name is not set.
This file will probably be used as a template for terraform.tfvars and the team_name needs to match the $TEAM_NAME environment variable which should be passed into the S3 bucket creation script. The only characters permitted in S3 bucket names are letters, numbers, dots and dashes. If we place a colon in the default value of team_name then the user must overtype it. If this is not done then the user will need to repeat some steps in the set-up.

Solo: @poveyd