Modify README and terraform/jenkins example config
For a consistent approach with an earlier step, change the example
config file to specify which settings the user should and shouldn't
change in a more human-readable way.

Add in comments above the lines that need to be changed to explain
what each setting is or does.

Change the README guidance to reflect these changes.