Move a configuration parameter higher in the configuration file
It happened to me that I missed to change the `team_name` variable
in the terraform configuration file, because it was at the bottom
after a variable that did not need to change (so I thought I was done).
The provisioning of the environment failed and it took me a while to understand why.
I have moved the variable to a more prominent place.