Complete example of full DNS and Jenkins Module Use
This example deploys a teams dns and jenkins instance within a single terraform apply.

If used by an team wishing to use the build.gds-reliability.engineering domain it is neccessary for the RE Build team to delegate DNS resolution to allow the ACM certificate to be DNS validated and issued.

Solo: @smford