Remove some examples, as they still need to be review
Some of the examples haven't been checked against the tech writer guidelines
and not yet been fully tested after the new way of modularising our Terraform code.

We are going to remove them for now until we have time to go through them.

I have checked this with Dave Povey, and he is OK about that.

A card in the backlog has been created:
https://trello.com/c/hqzhmgsp/225-review-other-examples-that-we-have-temporarily-removed